requirements.txt
Pillow
matplotlib
scipy
tensorflow